backend/requirements.txt

fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
python-dotenv
openai>=1.0.0
google-generativeai
google-generativeai